+++ docs/identifier_lookup/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this gear are documented in this file.
 
+## 1.0.3
+* Updates identifier lookup error messages and identifier file suffix.
+
 ## 1.0.0
 
 * Adds ability to do a reverse lookup on NACCID to find the center IDs.